New folder/LFA/farmer_chatbot_module/requirements.txt
# requirements.txt

# For connecting to the Google Generative AI (Gemini) API
google-generativeai

# For loading secret keys from the .env file
python-dotenv

# For making HTTP requests to the ML model APIs
requests